Extract unified download utility with progress bar
