+++ crates/hipparcos/Cargo.toml
@@ -8,10 +8,8 @@ repository.workspace = true
 
 [dependencies]
 starfield = { workspace = true }
-reqwest = { workspace = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
-indicatif = { workspace = true }
 rand = { workspace = true }
 nalgebra = { workspace = true }
 flate2.workspace = true
