+++ crates/hipparcos/src/downloader.rs
@@ -8,11 +8,7 @@ use std::path::{Path, PathBuf};
 
 use starfield::Result;
 use starfield::StarfieldError;
-use starfield_datasource_utils::{
-    build_http_client, check_response_status, ensure_cache_dir, file_exists_and_not_empty,
-};
-
-use indicatif::{ProgressBar, ProgressStyle};
+use starfield_datasource_utils::{download_to_file, ensure_cache_dir, file_exists_and_not_empty};
 
 // Hipparcos catalog URL
 const HIPPARCOS_URL: &str = "https://cdsarc.cds.unistra.fr/ftp/cats/I/239/hip_main.dat";
@@ -29,52 +25,6 @@ pub fn get_cache_dir() -> PathBuf {
     starfield_datasource_utils::cache_dir()
 }
 
-/// Download a file from URL to a local path
-fn download_file<P: AsRef<Path>>(url: &str, path: P) -> Result<()> {
-    // Create parent directories if they don't exist
-    if let Some(parent) = path.as_ref().parent() {
-        fs::create_dir_all(parent).map_err(StarfieldError::IoError)?;
-    }
-
-    // Create a temporary file first to avoid partial downloads
-    let temp_path = path.as_ref().with_extension("tmp");
-    let mut file = BufWriter::new(File::create(&temp_path).map_err(StarfieldError::IoError)?);
-
-    let client = build_http_client(30)?;
-
-    // Make the request
-    let mut response = check_response_status(
-        client
-            .get(url)
-            .send()
-            .map_err(|e| StarfieldError::DataError(format!("Failed to download file: {}", e)))?,
-        url,
-    )?;
-
-    // Copy the response body to the file
-    let mut buffer = [0; 8192];
-    loop {
-        let bytes_read = response
-            .read(&mut buffer)
-            .map_err(|e| StarfieldError::DataError(format!("Failed to read response: {}", e)))?;
-
-        if bytes_read == 0 {
-            break;
-        }
-
-        file.write_all(&buffer[..bytes_read])
-            .map_err(StarfieldError::IoError)?;
-    }
-
-    // Flush and sync the file
-    file.flush().map_err(StarfieldError::IoError)?;
-    drop(file);
-
-    // Rename the temporary file to the final path
-    fs::rename(temp_path, path).map_err(StarfieldError::IoError)?;
-
-    Ok(())
-}
 
 /// Decompress a gzipped file
 /// Currently unused as we're using synthetic data, but kept for future reference
@@ -153,75 +103,6 @@ pub fn resolve_url(filename: &str) -> Option<String> {
     None
 }
 
-/// Download a file from URL to a local path, showing a progress bar.
-///
-/// Uses a longer timeout (600s) suitable for large ephemeris files.
-/// Downloads to a temporary file first, then atomically renames.
-pub fn download_file_with_progress<P: AsRef<Path>>(url: &str, path: P) -> Result<()> {
-    if let Some(parent) = path.as_ref().parent() {
-        fs::create_dir_all(parent).map_err(StarfieldError::IoError)?;
-    }
-
-    let temp_path = path.as_ref().with_extension("tmp");
-    let mut file = BufWriter::new(File::create(&temp_path).map_err(StarfieldError::IoError)?);
-
-    let client = build_http_client(600)?;
-
-    let response = check_response_status(
-        client
-            .get(url)
-            .send()
-            .map_err(|e| StarfieldError::DataError(format!("Failed to download {}: {}", url, e)))?,
-        url,
-    )?;
-
-    let total_size = response.content_length().unwrap_or(0);
-    let pb = if total_size > 0 {
-        let pb = ProgressBar::new(total_size);
-        pb.set_style(
-            ProgressStyle::default_bar()
-                .template("[{bar:40}] {percent}% {bytes}/{total_bytes} ({bytes_per_sec})")
-                .unwrap()
-                .progress_chars("##-"),
-        );
-        Some(pb)
-    } else {
-        None
-    };
-
-    let mut reader = io::BufReader::new(response);
-    let mut downloaded: u64 = 0;
-    let mut buffer = [0u8; 131_072]; // 128KB chunks
-
-    loop {
-        let bytes_read = reader
-            .read(&mut buffer)
-            .map_err(|e| StarfieldError::DataError(format!("Failed to read response: {}", e)))?;
-
-        if bytes_read == 0 {
-            break;
-        }
-
-        file.write_all(&buffer[..bytes_read])
-            .map_err(StarfieldError::IoError)?;
-
-        downloaded += bytes_read as u64;
-        if let Some(ref pb) = pb {
-            pb.set_position(downloaded);
-        }
-    }
-
-    if let Some(ref pb) = pb {
-        pb.finish_and_clear();
-    }
-
-    file.flush().map_err(StarfieldError::IoError)?;
-    drop(file);
-
-    fs::rename(temp_path, path).map_err(StarfieldError::IoError)?;
-
-    Ok(())
-}
 
 /// Ensure a data file is available locally, downloading it if necessary.
 ///
@@ -251,7 +132,7 @@ pub fn download_or_cache(filename: &str, data_dir: Option<&Path>) -> Result<Path
     })?;
 
     eprintln!("Downloading {} ...", url);
-    download_file_with_progress(&url, &local_path)?;
+    download_to_file(&url, &local_path, 600)?;
     eprintln!("Saved to {}", local_path.display());
 
     Ok(local_path)
@@ -289,7 +170,7 @@ pub fn download_hipparcos() -> Result<PathBuf> {
     println!("This may take a moment as the catalog is approximately 36MB");
 
     // Attempt to download the file
-    match download_file(HIPPARCOS_URL, &dat_path) {
+    match download_to_file(HIPPARCOS_URL, &dat_path, 30) {
         Ok(_) => {
             println!(
                 "Hipparcos catalog downloaded successfully to {}",
@@ -310,6 +191,7 @@ pub fn download_hipparcos() -> Result<PathBuf> {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use starfield_datasource_utils::build_http_client;
 
     #[test]
     fn test_cache_dir() {
